Add files via upload
This is a U-food Marketing Data Analysis Project. In this project we have a data set with that contains customer information and the interaction of those customer with U-food Marketing Campaigns. In this case study we have done a through EDA and derived insights on customer demographics, business opportunities and data driven action to optimize marketing campaigns using python libraries.